@@ -9,32 +9,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "4.3.6",
-    "@angular/cdk": "github:angular/cdk-builds",
-    "@angular/common": "4.3.6",
-    "@angular/compiler": "4.3.6",
-    "@angular/core": "4.3.6",
-    "@angular/forms": "4.3.6",
-    "@angular/http": "4.3.6",
-    "@angular/material": "github:angular/material2-builds",
-    "@angular/platform-browser": "4.3.6",
-    "@angular/platform-browser-dynamic": "4.3.6",
-    "@angular/router": "4.3.6",
-    "bootstrap": "3.3.7",
-    "core-js": "2.5.0",
+    "@angular/animations": "5.2.9",
+    "@angular/cdk": "5.2.4",
+    "@angular/common": "5.2.9",
+    "@angular/compiler": "5.2.9",
+    "@angular/core": "5.2.9",
+    "@angular/flex-layout": "5.0.0-beta.13",
+    "@angular/forms": "5.2.9",
+    "@angular/material": "5.2.4",
+    "@angular/platform-browser": "5.2.9",
+    "@angular/platform-browser-dynamic": "5.2.9",
+    "@angular/router": "5.2.9",
+    "core-js": "2.5.3",
     "hammerjs": "2.0.8",
-    "rxjs": "5.4.3",
+    "npm": "^5.8.0",
+    "rxjs": "5.5.7",
     "web-animations-js": "2.3.1",
-    "zone.js": "0.8.17"
+    "zone.js": "0.8.20"
   },
   "devDependencies": {
-    "@angular/cli": "1.3.2",
-    "@angular/compiler-cli": "4.3.6",
-    "@types/node": "8.0.25",
-    "codelyzer": "3.1.2",
-    "ts-node": "3.3.0",
-    "tslint": "5.7.0",
-    "tslint-language-service": "0.9.6",
-    "typescript": "2.4.2"
+    "@angular/cli": "1.7.3",
+    "@angular/compiler-cli": "5.2.9",
+    "@types/node": "9.6.0",
+    "codelyzer": "4.2.1",
+    "ts-node": "5.0.1",
+    "tslint": "5.9.1",
+    "typescript": "2.5.3"
   }
 }
