@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 
-import { UserService } from './shared';
+import { UserService } from './shared/services/user.service';
 
 @Component({
   selector: 'app-root',
