@@ -1,16 +1,11 @@
-import {
-  Directive,
-  Input,
-  OnInit,
-  TemplateRef,
-  ViewContainerRef
-} from '@angular/core';
+import { Directive, Input, OnInit, TemplateRef, ViewContainerRef } from '@angular/core';
 
 import { UserService } from '../services/user.service';
+import { User } from '../types/interfaces';
 
 @Directive({ selector: '[showAuthed]' })
 export class ShowAuthedDirective implements OnInit {
-  condition: boolean;
+  @Input() showAuthed: boolean;
 
   constructor(
     private templateRef: TemplateRef<any>,
@@ -18,20 +13,13 @@ export class ShowAuthedDirective implements OnInit {
     private viewContainer: ViewContainerRef
   ) {}
 
-  @Input()
-  set showAuthed(condition: boolean) {
-    this.condition = condition;
-  }
-
   ngOnInit(): void {
-    this.userService.isAuthenticated.subscribe(
-      (isAuthenticated: boolean) => {
-        if (isAuthenticated && this.condition || !isAuthenticated && !this.condition) {
-          this.viewContainer.createEmbeddedView(this.templateRef);
-        } else {
-          this.viewContainer.clear();
-        }
+    this.userService.currentUser.subscribe((user: User) => {
+      if (user != null && this.showAuthed || user == null && !this.showAuthed) {
+        this.viewContainer.createEmbeddedView(this.templateRef);
+      } else {
+        this.viewContainer.clear();
       }
-    );
+    });
   }
 }
