@@ -1,48 +1,26 @@
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Headers, Http, Response, URLSearchParams } from '@angular/http';
 import { Observable } from 'rxjs/Rx';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
 
 @Injectable()
 export class ApiService {
   constructor(
-    private http: Http
+    private http: HttpClient
   ) {}
 
-  get(path: string, params: URLSearchParams = new URLSearchParams()): Observable<any> {
-    return this.http.get(`/api${path}`, { headers: this.setHeaders(), search: params })
-      .catch(this.formatErrors)
-      .map((res: Response) => res.json());
+  get(path: string): Observable<any> {
+    return this.http.get(`/api${path}`);
   }
 
   post(path: string, body: any = {}): Observable<any> {
-    return this.http.post(`/api${path}`, JSON.stringify(body), { headers: this.setHeaders() })
-      .catch(this.formatErrors)
-      .map((res: Response) =>  res.json());
+    return this.http.post(`/api${path}`, body);
   }
 
   put(path: string, body: any = {}): Observable<any> {
-    return this.http.put(`/api${path}`, JSON.stringify(body), { headers: this.setHeaders() })
-      .catch(this.formatErrors)
-      .map((res: Response) => res.json());
+    return this.http.put(`/api${path}`, body);
   }
 
   delete(path: string): Observable<any> {
-    return this.http.delete(`/api${path}`, { headers: this.setHeaders() })
-      .catch(this.formatErrors)
-      .map((res: Response) => res.json());
-  }
-
-  private setHeaders(): Headers {
-    const headersConfig = {
-      'Content-Type': 'application/json',
-      Accept: 'application/json'
-    };
-    return new Headers(headersConfig);
-  }
-
-  private formatErrors(error: any): Observable<any> {
-    return Observable.throw(error.json());
+    return this.http.delete(`/api${path}`);
   }
 }
